Split Jira issue workspace presentation
